Add exynos support
Conflicts:
	include/gui/SensorManager.h
	libs/gui/SensorManager.cpp

Conflicts:
	include/gui/SensorManager.h
	libs/binder/Android.mk
	libs/binder/IMemory.cpp
	libs/gui/SensorManager.cpp
	services/surfaceflinger/SurfaceFlinger.cpp
	services/surfaceflinger/SurfaceFlinger.h

+++ libs/ui/GraphicBuffer.cpp
@@ -197,7 +197,7 @@ status_t GraphicBuffer::lock(uint32_t inUsage, const Rect& rect, void** vaddr)
                 width, height);
         return BAD_VALUE;
     }
-    status_t res = getBufferMapper().lock(handle, inUsage, rect, vaddr);
+    status_t res = getBufferMapper().lock(handle, static_cast<int>(inUsage), rect, vaddr);
     return res;
 }
 
@@ -218,7 +218,7 @@ status_t GraphicBuffer::lockYCbCr(uint32_t inUsage, const Rect& rect,
                 width, height);
         return BAD_VALUE;
     }
-    status_t res = getBufferMapper().lockYCbCr(handle, inUsage, rect, ycbcr);
+    status_t res = getBufferMapper().lockYCbCr(handle, static_cast<int>(inUsage), rect, ycbcr);
     return res;
 }
 
@@ -245,7 +245,7 @@ status_t GraphicBuffer::lockAsync(uint32_t inUsage, const Rect& rect,
                 width, height);
         return BAD_VALUE;
     }
-    status_t res = getBufferMapper().lockAsync(handle, inUsage, rect, vaddr,
+    status_t res = getBufferMapper().lockAsync(handle, static_cast<int>(inUsage), rect, vaddr,
             fenceFd);
     return res;
 }
@@ -268,7 +268,7 @@ status_t GraphicBuffer::lockAsyncYCbCr(uint32_t inUsage, const Rect& rect,
                 width, height);
         return BAD_VALUE;
     }
-    status_t res = getBufferMapper().lockAsyncYCbCr(handle, inUsage, rect,
+    status_t res = getBufferMapper().lockAsyncYCbCr(handle, static_cast<int>(inUsage), rect,
             ycbcr, fenceFd);
     return res;
 }
